+++ database/seeds/DatabaseSeeder.php
@@ -14,6 +14,7 @@ class DatabaseSeeder extends Seeder
         App\User::create(['name' => "JP Murray", "email" => "curieuxmurray@gmail.com", "password" => Illuminate\Support\Facades\Hash::make('secret')]);
         $this->call(TaxonomySeeder::class);
         $this->call(SpeciesTableSeeder::class);
+        $this->call(SupplierSeeder::class);
         // $this->call(UsersTableSeeder::class);
     }
 }