+++ resources/views/species/show.blade.php
@@ -24,6 +24,13 @@
 			        <p class="card-text">
 			        	FR: {{ $species->name_fr }}<br>
 			        	EN: {{ $species->name_en }}<br>
+			        </p>
+			        @if(!empty($species->interesting_cultivar))
+			        <p class="card-text">
+			        	<strong>Interesting cultivars:</strong> {{ implode(', ', $species->interesting_cultivar) }}
+			        </p>
+			        @endif
+			        <p class="card-text">
 			        	<small class="text-muted">Dernière mise à jour: {{ $species->updated_at->tz("America/Montreal") }}</small>
 			        </p>
 			      </div>
@@ -74,14 +81,10 @@
 				  </div>
 				  <div class="card-body">
 				    <p class="card-text"><strong>Comestible</strong>:
-				    	@if($species->comestible_use == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->comestible_use as $use)
 				    		{{ ucfirst($use) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				    <p class="card-text"><strong>Medicinal</strong>: {{ $species->medicinal_use ? "Yes" : "No" }}
 				    </p>
@@ -96,44 +99,28 @@
 				  </div>
 				  <div class="card-body">
 				    <p class="card-text"><strong>Flowering period</strong>:
-				    	@if($species->flowering_period == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->flowering_period as $period)
 				    		{{ ucfirst($period) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				    <p class="card-text"><strong>Flowering color</strong>:
-				    	@if($species->flowering_color == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->flowering_color as $color)
 				    		{{ ucfirst($color) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				    <p class="card-text"><strong>Foliage color</strong>:
-				    	@if($species->foliage_color == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->foliage_color as $color)
 				    		{{ ucfirst($color) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				    <p class="card-text"><strong>Post summer appeal</strong>:
-				    	@if($species->post_summer_appeal == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->post_summer_appeal as $season)
 				    		{{ ucfirst($season) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				  </div>
 				</div>
@@ -144,84 +131,93 @@
 				  </div>
 				  <div class="card-body">
 				    <p class="card-text"><strong>Growth</strong>:
-				    	@if($species->growth == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->growth as $speed)
 				    		{{ ucfirst($speed) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				    <p class="card-text"><strong>Pruning period</strong>:
-				    	@if($species->pruning_period == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->pruning_period as $period)
 				    		{{ ucfirst($period) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				    <p class="card-text"><strong>Multiplication</strong>:
-				    	@if($species->multiplication == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->multiplication as $method)
 				    		{{ ucfirst($method) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				    <p class="card-text"><strong>Disadvantages</strong>:
-				    	@if($species->disadvantages == null)
-				    	None or unknown.
-				    	@else
 				    	@foreach($species->disadvantages as $disadvantage)
 				    		{{ ucfirst($disadvantage) }}
 				    		@if(!$loop->last && $loop->count != 1) , @endif
 				    	@endforeach
-				    	@endif
 				    </p>
 				  </div>
 				</div>
 			</div>
 
-			<div class="card">
-			  <h5 class="card-header">Functions</h5>
-			  <div class="card-body">
-			    <p class="card-text"><strong>Nitrogen fixer</strong>:
-			    	{{ $species->nitrogen_fixer ? "Yes" : "No" }}
-			    </p>
-			    <p class="card-text"><strong>Nutrients accumulator</strong>:
-			    	{{ $species->nutrient_accumulator ? "Yes" : "No" }}
-			    </p>
-			    <p class="card-text"><strong>Ground cover</strong>:
-			    	{{ $species->ground_cover ? "Yes" : "No" }}
-			    </p>
-			    <p class="card-text"><strong>Hedge</strong>:
-			    	{{ $species->hedge ? "Yes" : "No" }}
-			    </p>
-			    <p class="card-text"><strong>Wildlife use</strong>:
-			    	@foreach($species->wildlife_use as $use)
-			    		{{ ucfirst($use) }}
-			    		@if(!$loop->last && $loop->count != 1) , @endif
-			    	@endforeach
-			    </p>
-			    <p class="card-text"><strong>Ecological use</strong>:
-			    	@foreach($species->ecological_use as $use)
-			    		{{ ucfirst($use) }}
-			    		@if(!$loop->last && $loop->count != 1) , @endif
-			    	@endforeach
-			    </p>
-			    <p class="card-text"><strong>Pollinating type</strong>:
-			    	@foreach($species->pollinating_type as $type)
-			    		{{ ucfirst($type) }}
-			    		@if(!$loop->last && $loop->count != 1) , @endif
-			    	@endforeach
-			    </p>
-			  </div>
+			<div class="card-deck mb-3">
+				<div class="card">
+			  		<h5 class="card-header">Functions</h5>
+
+					<div class="card-body">
+						<p class="card-text"><strong>Nitrogen fixer</strong>:
+							{{ $species->nitrogen_fixer ? "Yes" : "No" }}
+						</p>
+						<p class="card-text"><strong>Nutrients accumulator</strong>:
+							{{ $species->nutrient_accumulator ? "Yes" : "No" }}
+						</p>
+						<p class="card-text"><strong>Ground cover</strong>:
+							{{ $species->ground_cover ? "Yes" : "No" }}
+						</p>
+						<p class="card-text"><strong>Hedge</strong>:
+							{{ $species->hedge ? "Yes" : "No" }}
+						</p>
+						<p class="card-text"><strong>Wildlife use</strong>:
+							@foreach($species->wildlife_use as $use)
+								{{ ucfirst($use) }}
+								@if(!$loop->last && $loop->count != 1) , @endif
+							@endforeach
+						</p>
+						<p class="card-text"><strong>Ecological use</strong>:
+							@foreach($species->ecological_use as $use)
+								{{ ucfirst($use) }}
+								@if(!$loop->last && $loop->count != 1) , @endif
+							@endforeach
+						</p>
+						<p class="card-text"><strong>Pollinating type</strong>:
+							@foreach($species->pollinating_type as $type)
+								{{ ucfirst($type) }}
+								@if(!$loop->last && $loop->count != 1) , @endif
+							@endforeach
+						</p>
+					</div>
+				</div>
+
+				@if($species->suppliers->isNotEmpty())
+				<div class="card">
+			  		<h5 class="card-header">Suppliers</h5>
+
+					<div class="card-body">
+						<ul>
+							@foreach($species->suppliers as $supplier)
+							<li><a href="{{ $supplier->url }}" target="_blank">{{ $supplier->name }}</a></li>
+							@endforeach
+						</ul>
+					</div>
+				</div>
+				@endif
 			</div>
+
+			@if(!empty($species->maintainers_note))
+			<div class="card mb-3">
+			  <h5 class="card-header">Database maintainers' notes</h5>
+			  <div class="card-body">
+			    {!! $species->maintainers_note !!}
+			</div>
+			@endif
         </div>
     </div>
 </div>