+++ resources/views/suppliers/edit.blade.php
@@ -1,0 +1,62 @@
+@extends('layouts.app')
+
+@section('content')
+<div class="container">
+    <div class="row justify-content-center">
+        <div class="col-md-12">
+            @if (session('status'))
+                <div class="alert alert-success" role="alert">
+                    {{ session('status') }}
+                </div>
+            @endif
+            
+            <div class="card">
+                <div class="card-header">Edit {{ $supplier->name }}</div>
+
+                <div class="card-body">
+                    <form method="POST" action="{{ route('suppliers.update', $supplier) }}">
+                        @csrf
+                        @method('PUT')
+
+                        <div class="form-group row">
+                            <label for="name" class="col-md-3 col-form-label text-md-right">{{ __('Name') }}</label>
+
+                            <div class="col-md-6">
+                                <input id="name" type="text" class="form-control @error('name') is-invalid @enderror" name="name" value="{{ $supplier->name }}" required autocomplete="name" autofocus>
+
+                                @error('name')
+                                    <span class="invalid-feedback" role="alert">
+                                        <strong>{{ $message }}</strong>
+                                    </span>
+                                @enderror
+                            </div>
+                        </div>
+
+                        <div class="form-group row">
+                            <label for="url" class="col-md-3 col-form-label text-md-right">{{ __('URL') }}</label>
+
+                            <div class="col-md-6">
+                                <input id="url" type="text" class="form-control @error('url') is-invalid @enderror" name="url" value="{{ $supplier->url }}" required autocomplete="url" autofocus>
+
+                                @error('url')
+                                    <span class="invalid-feedback" role="alert">
+                                        <strong>{{ $message }}</strong>
+                                    </span>
+                                @enderror
+                            </div>
+                        </div>
+
+                        <div class="form-group row mb-0">
+                            <div class="col-md-6 offset-md-3">
+                                <button type="submit" class="btn btn-primary">
+                                    {{ __('Update') }}
+                                </button>
+                            </div>
+                        </div>
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+@endsection