+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+# 0.5.0 - Pre-release (2019-08-25)
+- Fixed some create / edit form quirkness.
+- Database fields changed for a couple of _species_ fields to better show current usage.
+- Ability to add interesting cultivars to a species.
+- Ability to add maintainers' notes to a species.
+- Icon to see public pages for a species in its edit page.
+- Summernote added for maintainer's notes.
+- Manage suppliers and associates them to species.
+
 # 0.4.0 - Pre-release (2019-08-25)
 - Fix comestible usage information on show page.
 - Changed layout of _species_ create and edit pages.